# scheduler/app/docs/swagger.yaml
basePath: /
definitions:
  main.NodeOutputChunkResponse:
    properties:
      data:
        type: string
      done:
        type: boolean
      items:
        items: {}
        type: array
      limit:
        type: integer
      mode:
        type: string
      next_offset:
        type: integer
      offset:
        type: integer
      total_chars:
        type: integer
      total_items:
        type: integer
    type: object
  main.ProcessPaymentsResponse:
    properties:
      processed_count:
        type: integer
    type: object
  main.RegisterWorkerRequest:
    properties:
      worker_id:
        type: string
    type: object
  main.adminWorkflowActivateRequest:
    properties:
      reset_state:
        type: boolean
      workflow_id:
        type: string
    type: object
  main.adminWorkflowActivateResponse:
    properties:
      enqueued_jobs:
        type: integer
      recovered_nodes:
        type: integer
      reset_state:
        type: boolean
      topological_size:
        type: integer
      workflow_id:
        type: string
    type: object
  main.adminWorkflowDeleteRequest:
    properties:
      workflow_id:
        type: string
    type: object
  main.adminWorkflowDeleteResponse:
    properties:
      deleted:
        type: boolean
      workflow_id:
        type: string
    type: object
  main.adminWorkflowListResponse:
    properties:
      active_workflow_id:
        type: string
      loaded_workflow_id:
        type: string
      uploaded_ids:
        items:
          type: string
        type: array
    type: object
  scheduler.Assignment:
    properties:
      args:
        items: {}
        type: array
      dependencies:
        items:
          $ref: '#/definitions/scheduler.DependencyRef'
        type: array
      job_id:
        type: string
      node_id:
        type: string
      required_rep:
        type: integer
      reward_usdc:
        type: string
      wasm_url:
        type: string
      workflow_id:
        type: string
    type: object
  scheduler.Decision:
    properties:
      accepted_payload:
        additionalProperties: {}
        type: object
      accepted_result:
        type: string
      accepted_workers:
        type: integer
      finalized:
        type: boolean
      job_id:
        type: string
      required_replicas:
        type: integer
    type: object
  scheduler.DependencyRef:
    properties:
      node_id:
        type: string
      workflow_id:
        type: string
    type: object
  scheduler.PaymentEvent:
    properties:
      accepted_hash:
        type: string
      amount_usdc:
        type: string
      attempts:
        type: integer
      id:
        type: string
      job_id:
        type: string
      last_error:
        type: string
      network:
        type: string
      payer:
        type: string
      status:
        type: string
      tx_hash:
        type: string
      updated_at:
        type: string
      worker_id:
        type: string
      workflow_id:
        type: string
    type: object
  scheduler.ResultSubmission:
    properties:
      job_id:
        type: string
      result_payload:
        additionalProperties: {}
        type: object
      result_sig:
        type: string
      worker_id:
        type: string
    type: object
  scheduler.Stats:
    properties:
      finalized_jobs:
        type: integer
      pending_payments:
        type: integer
      queued_jobs:
        type: integer
      total_jobs:
        type: integer
      workers_online:
        type: integer
    type: object
  scheduler.Worker:
    properties:
      connected_at:
        type: string
      id:
        type: string
      last_heartbeat:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Worker registration, job pull, result submit, payments, stats.
  title: x402 Scheduler API
  version: "1.0"
paths:
  /api/admin/workflows:
    get:
      parameters:
      - description: Bearer admin token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.adminWorkflowListResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: List workflows and active workflow id
      tags:
      - admin
  /api/admin/workflows/activate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer admin token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Activate request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.adminWorkflowActivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.adminWorkflowActivateResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Activate workflow by id
      tags:
      - admin
  /api/admin/workflows/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer admin token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Delete request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.adminWorkflowDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.adminWorkflowDeleteResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Delete uploaded workflow and its persisted state
      tags:
      - admin
  /api/admin/workflows/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer admin token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Workflow JSON file
        in: formData
        name: workflow_json
        required: true
        type: file
      - description: One or more C++ files (.cpp)
        in: formData
        name: cpp_files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Upload workflow JSON and C++ sources
      tags:
      - admin
  /api/payments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scheduler.PaymentEvent'
            type: array
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: List payment queue
      tags:
      - payments
  /api/payments/process:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ProcessPaymentsResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Process pending payments
      tags:
      - payments
  /api/pull:
    get:
      parameters:
      - description: Worker ID
        in: query
        name: worker_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduler.Assignment'
        "204":
          description: No assignment
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Pull next assignment
      tags:
      - workers
  /api/result:
    post:
      consumes:
      - application/json
      parameters:
      - description: Result payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/scheduler.ResultSubmission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduler.Decision'
        "400":
          description: Bad Request
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Submit job result
      tags:
      - jobs
  /api/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduler.Stats'
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Scheduler stats snapshot
      tags:
      - stats
  /api/workers/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Worker register request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.RegisterWorkerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduler.Worker'
        "400":
          description: Bad Request
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Register worker or heartbeat
      tags:
      - workers
  /api/workflow/node-output:
    get:
      parameters:
      - description: Workflow ID
        in: query
        name: workflow_id
        required: true
        type: string
      - description: Node ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Get finalized workflow node output
      tags:
      - workflow
  /api/workflow/node-output/chunk:
    get:
      parameters:
      - description: Workflow ID
        in: query
        name: workflow_id
        required: true
        type: string
      - description: Node ID
        in: query
        name: node_id
        required: true
        type: string
      - description: Chunk offset
        in: query
        name: offset
        type: integer
      - description: Chunk size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.NodeOutputChunkResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
      summary: Get chunk of finalized workflow node output
      tags:
      - workflow
swagger: "2.0"